Move dependent files to namespace
