+++ osu.Game/Graphics/UserInterfaceV2/FileSelection/OsuDirectorySelectorDirectory.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics.Sprites;
 
-namespace osu.Game.Graphics.UserInterfaceV2
+namespace osu.Game.Graphics.UserInterfaceV2.FileSelection
 {
     internal partial class OsuDirectorySelectorDirectory : DirectorySelectorDirectory
     {
@@ -23,7 +23,7 @@ namespace osu.Game.Graphics.UserInterfaceV2
             Flow.AutoSizeAxes = Axes.X;
             Flow.Height = OsuDirectorySelector.ITEM_HEIGHT;
 
-            AddInternal(new OsuFileSelectorBackgroundLayer());
+            AddInternal(new BackgroundLayer());
 
             Colour = colours.Orange1;
         }
