+++ osu.Game/Graphics/UserInterfaceV2/OsuDirectorySelector.cs
@@ -8,6 +8,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics.Containers;
+using osu.Game.Graphics.UserInterfaceV2.FileSelection;
 using osu.Game.Overlays;
 
 namespace osu.Game.Graphics.UserInterfaceV2
@@ -57,7 +58,7 @@ namespace osu.Game.Graphics.UserInterfaceV2
                 {
                     RelativeSizeAxes = Axes.Both,
                 },
-                new OsuDirectorySelectorHiddenToggle
+                new HiddenFilesToggleCheckbox
                 {
                     Current = { BindTarget = ShowHiddenItems },
                 },
